Update COLR tests
Running gen-shaping-tests.py add a number of tests that wasn’t added
before. I don’t know the history of this, but some of the tests pass and
others don’t, so I have added the ones that do not pass to the list of
ignored tests and reran the script.