Check coverage before reading subtables
The intent is to avoid both redundant and unnecessary work. For each subtable, before doing the full parse, read the primary coverage table and then compute the index for the current glyph. If successful, bundle up the glyph, coverage table and coverage index so we can pass that data on to the subtable apply function. If unsuccessful, don't parse the subtable at all.