Drop --show-extents from HarfBuzz tests (#286)
We were already disabling most of such extents since the do not match
Skrifa extents. Since glyphs extents are not interesting for shaping,
and we are effectively testing Skrifa code here not HarfRust, we can
skip testing extents entirely.

We could even drop such tests entirely, but I think it does not harm to
test only shaping for these.

This allows us to remove most of the remaining tests in the ignored
tests list.